Revamp the word language model example
Notable changes:

- Boiled down predict.py to a point where, if there are no
  further additions I would argue that it could be removed.
- Removed optimizer parameter in train_step
- Removed a lot of FIXME/TODO which solved themselves over time
- Removed statefulness of predict/validation
- Add manual seeding
