+++ examples/word_language_model/predict.py
@@ -6,36 +6,26 @@ from torch.autograd import Variable
 
 import data
 import model
-import learner
+import net
 
 parser = argparse.ArgumentParser(description='PyTorch PennTreeBank RNN/LSTM Language Model')
 parser.add_argument('--data', type=str, default='./data/penn',
                     help='location of the data corpus')
-parser.add_argument('--bptt', type=int, default=35,
-                    help='sequence length')
 parser.add_argument('--seed', type=int, default=1111,
                     help='random seed')
 parser.add_argument('--no-cuda', dest='cuda', action='store_false',
                     help='use CUDA')
 parser.add_argument('--checkpoint', type=str, default='./model.pt',
                     help='model checkpoint to use')
-parser.add_argument('--outf', type=str, default='generated.txt',
-                    help='output file for generated text')
-parser.add_argument('--temperature', type=float, default=1.0,
-                    help='temperature - higher will increase diversity')
-parser.add_argument('--words', type=int, default='1000',
-                    help='number of words to generate')
-parser.add_argument('--log-interval', type=int, default=100,
-                    help='reporting interval')
 
 args = parser.parse_args()
 
-# TODO: set seed
+torch.manual_seed(args.seed)
 
 corpus = data.Corpus(args.data)
 ntokens = len(corpus.dictionary)
 
-learner = learner.Learner(
+net = net.Net(
     module=model.RNNModel,
     batch_size=1,
     use_cuda=args.cuda,
@@ -44,25 +34,21 @@ learner = learner.Learner(
     module__ninp=200,
     module__nhid=200,
     module__nlayers=2)
-learner.initialize()
-learner.load_params(args.checkpoint)
+net.initialize()
+net.load_params(args.checkpoint)
 
-words = [corpus.dictionary.idx2word[n] for n in range(10)]
 
-print(words)
-
-p = learner.predict_proba(np.array([[
+input_data = np.array([[
     corpus.dictionary.word2idx['fish'],
     corpus.dictionary.word2idx['sees'],
     corpus.dictionary.word2idx['man'],
-]]))
+]])
 
+# Convert to batch last, i.e. (b, t) -> (t, b).
+input_data = input_data.T
+
+p = net.predict_proba(input_data)
 print(p)
 
-widx = learner.predict(np.array([[
-    corpus.dictionary.word2idx['fish'],
-    corpus.dictionary.word2idx['sees'],
-    corpus.dictionary.word2idx['man'],
-]]))
-
+widx = net.predict(input_data)
 print(widx, [corpus.dictionary.idx2word[n] for n in widx])
